// packages/horizon-intl/src/types/interfaces.ts
export interface TokenContext {
  // token 索引值的偏移量
  offset: number;

  // token 开始计算的初始行号
  line: number;

  // token 开始计算的初始列号
  col: number;

  // 原始输入
  text: string;

  // 换行数
  lineNum: number;
}

export interface Content {
  type: 'content';
  value: string;
  ctx: TokenContext;
}

// 需要解析参数定义
export interface PlainArg {
  type: 'argument';
  arg: string;
  ctx: TokenContext;
}

export interface Octothorpe {
  type: 'octothorpe';
  ctx: TokenContext;
}

export interface FunctionArg {
  type: 'function';
  arg: string;
  key: string;
  param?: Array<Content | PlainArg | FunctionArg | Select | Octothorpe>;
  ctx: TokenContext;
}

export interface SelectCase {
  key: string;
  tokens: Array<Content | PlainArg | FunctionArg | Select | Octothorpe>;
  ctx: TokenContext;
}

// 选择模式
export interface Select {
  type: 'plural' | 'select' | 'selectordinal';
  arg: string;
  cases: SelectCase[];
  offset?: number;
  ctx: TokenContext;
}